WriteBase里面，写Int16时的一个BUG

+++ NewLife.Core/Serialization/Base/WriterBase.cs
@@ -266,7 +266,7 @@ namespace NewLife.Serialization
                     Write((Byte)0);
                     return true;
                 case TypeCode.Int16:
-                    Write((Int32)value);
+                    Write((Int16)value);
                     return true;
                 case TypeCode.Int32:
                     Write((Int32)value);
